+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.1.0-rc.3 - 13 Dec 2022
+Change:
+- Minimum requirement for Elysia is set to 0.1.0-rc.5
+- Using `PreContext` instead of `Context`
+
+Fix:
+- Headers not present on Error
+
 # 0.1.0-rc.2 - 9 Dec 2022
 Fix:
 - Add main fields Bundlephobia